Exercise 4.21

+++ part4/blog/controllers/blog.js
@@ -43,8 +43,29 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  
+  const user = await User.findById(decodedToken.id)
+   
+  if (!user) {
+    return response.status(400).json({ error: 'userId missing or not valid' })
+  }
+  
+  const blog = await Blog.findById(request.params.id)
+  
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found (missing id or invalid id)' })
+  }
+  
+  if (blog.user.toString() === user.id) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'unauthorized delete operation' })
+  }
 })
 
 blogRouter.put('/:id', async (request, response) => {
